refactor(st-09029): split plan-execute node responsibilities
