+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-04-17
-**Active Story:** ST-09029 (Ready)
+**Last Updated:** 2026-04-18
+**Active Story:** ST-09029 (In Progress)
 
 ---
